Move BsplineTrajectory to common/trajectories
